SCRUM-1110 replaced old version of sort_dqm_json_reference_updates.py with new one using variable switch for api vs sqlalchemy batch
